corrected migrations

+++ db/migrate/20120309165530_create_games.rb
@@ -1,14 +1,14 @@
 class CreateGames < ActiveRecord::Migration
   def change
     create_table :games do |t|
-      t.int :round
+      t.integer :round
       t.datetime :date
-      t.int :stadium_id
-      t.int :home_id
-      t.int :visitor_id
-      t.int :home_score
-      t.int :visitor_score
-      t.double :attendance
+      t.integer :stadium_id
+      t.integer :home_id
+      t.integer :visitor_id
+      t.integer :home_score
+      t.integer :visitor_score
+      t.float :attendance
       t.float :income
       t.string :url
